Add maintenance script to manually migrate all or specific users (#3)
This commit adds a maintenance script that allows administrators to manually migrate users to WSOAuth. Unlike the `$wgOAuthMigrateUsersByUsername` setting, where a user will be migrated when they log in, this script forcefully migrates the specified user to WSOAuth, regardless of whether they have logged in or what the `$wgOAuthMigrateUsersByUsername` setting is.

This maintenance script aims to mitigate some of the effects version 3.0 has on existing installations, since it allows administrators to migrate users manually, without having to set `$wgOAuthMigrateUsersByUsername`.