Add Github provider
This is a copy of the Facebook provider, with more or less a simple
search/replace on Facebook -> Github.

The "name" is the Github Nickname, as otherwise it will show a
numeric name. This Nickname can be changed in Github by the user.

Realname and email are not implemented, as it requires more "scope".